+++ CHANGELOG.md
@@ -6,13 +6,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.2.0] - 2017-10-18
+### Added
+* Support arc command
+
 ## [1.1.1] - 2017-10-17
-* Bezier Curves
+### Added
+* Support Bézier curve commands
 * Allow negative / floating point numbers
 
 ## [1.0.2] - 2017-10-17
-* Basic moveTo, lineTo, and close commands
+### Added
+* Support basic moveTo, lineTo, and close commands
 
-[Unreleased]: https://github.com/ksm2/svg-parse/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/ksm2/svg-parse/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/ksm2/svg-parse/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/ksm2/svg-parse/compare/v1.0.2...v1.1.1
 [1.0.2]: https://github.com/ksm2/svg-parse/compare/10ebde2e0dea2845bada8418f8d8b835a54fab08..v1.0.2